Review edit actions, move to trait
