+++ controllers/users/_hint_trashed.php
@@ -2,8 +2,12 @@
     <div class="callout callout-danger">
         <div class="header">
             <i class="icon-minus-circle"></i>
-            <h3><?= __("User has deactivated their account") ?></h3>
-            <p><?= __("This user has deactivated their account and no longer wants to appear on the site. They can restore their account at any time by signing back in.") ?></p>
+            <h3><?= __("User account is deactivated") ?></h3>
+            <p>
+                <?= __("This user account has been deactivated can no longer be used.") ?>
+                <?= __("This account cannot be deleted since it contains internal references.") ?>
+                <?= __("To remove any personal data for this account, click the Anonymize Data button.") ?>
+            </p>
         </div>
     </div>
 </div>
